refactor(tui): extract terminal input into tui::input
Key classification and PTY byte encodings, xterm mouse encoding, and the
event-pump/frame pacing helpers each get their own file, with the tests
that cover them. Adds tui::test_support for fixtures more than one
submodule needs.
