+++ src/tui/input/mod.rs
@@ -1,0 +1,5 @@
+//! Terminal input handling: event intake, key encoding, mouse encoding.
+
+pub(in crate::tui) mod events;
+pub(in crate::tui) mod keys;
+pub(in crate::tui) mod mouse;
